Sync schedules between hubs in a group
Note there are no events that notify when hub schedules change, so we do
the sync when a player closes a hub GUI. If a script modifies the
schedule, we will miss it.

This would be so much simpler if Wube let space platforms use the group
feature on schedules.