Fix more schedule issues
Setting a record as active with go_to_station() causes a state change
event, even though the state does not change, and resets the platform's
spot in the queue (not sure why).

By instead having the active "fall down" into the record we want, as
long as it is the same plannet (which it always will be), no state
change happens.