Add search functionality to change group window
The version required change is the first version that has
helpers.compare_versions(). There is no way to safely use
helpers.multilingual_to_lower() before this version, as it was added in
2.0.67.

If the logistics bug is fixed, we will either require that version, or
do another compare_versions() check to only show providing limits if
that version.