Rework platform location queuing
Rather than maintain our own queue data structure tracking when
platforms arrived at planets, rely on the game's tracking in the
schedule object.

This simplifies quite a bit of tracking, and reduces potential points of
drift.

It also means we only need to calculate queue information when a
platform changes state, as we cache the platform queue indexes for use
in the GUI.

Also, change the logic to not include paused platforms in limitation.